Raising the Ante
Added Raising the Ante.  I thought that this could be used to boost the attack of Toons instead of Rolled, which is a Cog attack in Corporate Clash.
Pushing to this branch because if anything goes wrong, at least it's not on the goodbranch and because I don't know if there are changes in progress on the goodbranch.