Designated Attack Taunts For Cogs
Added functionality for certain Cogs to say certain things without the need for creating a new attack!  This functionality is currently used on the Litigation Team.

This functionality is somewhat bugged.  There's no real errors, but there are certain cases where a Cog will not say its unique taunt and say something else due to cheat definitions not using the standard phrase functionality, and the Case Manager will also not say his unique phrases due to it being overridden by an if/else.  I was not feeling like tweaking those at that moment.