Attack Dialogue Improvements
Most attacks that go into cheats have had their attack dialogue improved.  Instead of a hardcoded list to randomize, it uses the same method to get the Cog's attack dialogue that is used in getSuitTrack() and getSuitAnimTrack(), which can make a Cog with unique phrases say those instead of the generic dialogue.