Backticks to repr()
Changed all backtick expressions to instead use the repr() method.  Backticks were a form of syntax used until Python 3.x, which is when it was retired.  While it never caused any errors or crashes (due to us using a version of Panda3D with an outdated Python), and this shouldn't change the way the code functions, I believe it is good to bring the syntax up-to-date and so that these errors can be silenced.