Cog Shorthand Expansion
Expanded the maximum limit for a Cog shorthand from three characters to 20.  To reflect this, both the Prethinker and Witch Hunter's shorthands have been changed to reflect their shorthands in Toontown: Corporate Clash, with more to come (I didn't want to do all at once, especially not the Litigation Team, since I did not want it to be intrusive or confuse others).